#84 refactor: ProductDetail 관련 DTO package 이동
