+++ src/main/java/com/webeye/backend/productdetail/application/ProductDetailService.java
@@ -3,12 +3,12 @@ package com.webeye.backend.productdetail.application;
 import com.webeye.backend.imageanalysis.infrastructure.ImageUrlExtractor;
 import com.webeye.backend.imageanalysis.infrastructure.OpenAiClient;
 import com.webeye.backend.product.domain.Product;
-import com.webeye.backend.product.dto.request.ProductDetailAnalysisRequest;
-import com.webeye.backend.product.dto.response.DetailExplanationResponse;
+import com.webeye.backend.productdetail.dto.request.ProductDetailAnalysisRequest;
+import com.webeye.backend.productdetail.dto.response.DetailExplanationResponse;
 import com.webeye.backend.product.persistent.ProductRepository;
 import com.webeye.backend.productdetail.domain.ProductDetail;
 import com.webeye.backend.productdetail.domain.type.OutlineType;
-import com.webeye.backend.productdetail.dto.AllDetailExplanationResponse;
+import com.webeye.backend.productdetail.dto.response.AllDetailExplanationResponse;
 import com.webeye.backend.productdetail.persistent.ProductDetailRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
